Finish VisualCommitButton
